server/requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
django-cors-headers>=3.11.0
djangorestframework>=3.13.1
Pillow
boto3>=1.20.37